move Epmc step to Literature remove epmc job from workflow move Epmc config to literature section
