+++ src/main/resources/reference.conf
@@ -77,70 +77,6 @@ common {
 
 }
 
-epmc {
-  uris {
-    ensembl = "https://www.ensembl.org/Homo_sapiens/Gene/Summary?db=core;g="
-    chembl = "https://www.ebi.ac.uk/chembl/compound_report_card/"
-    ontologies = "https://www.ebi.ac.uk/ols/ontologies/efo/terms?short_form="
-  }
-  input {
-    cooccurences = ${literature.processing.outputs.cooccurrences}
-  }
-  outputs {
-    epmc-cooccurrences {
-      format = ${common.output-format}
-      path = ${common.output}"/literature/epmcCooccurrences"
-    }
-    output {
-      format = "json" // to be consistent with the other Evidence files
-      path = ${common.input}"/evidence-files/epmc"
-      options = [
-        {k: "compression", v: "gzip"},
-      ]
-    }
-  }
-  excluded-target-terms = [
-    "TEC",
-    "TECS",
-    "Tec",
-    "tec",
-    "'",
-    "(",
-    ")",
-    "-",
-    "-S",
-    "S",
-    "S-",
-    "SS",
-    "SSS",
-    "Ss",
-    "Ss-",
-    "s",
-    "s-",
-    "ss",
-    "U3",
-    "U6",
-    "u6",
-    "SNORA70",
-    "U2",
-    "U8"
-  ]
-  # The following are the relevant sections for disease/target associations as described in PMID28587637
-  sections-of-interest = [
-    "title",
-    "abstract",
-    "intro",
-    "case",
-    "figure",
-    "table",
-    "discuss",
-    "concl",
-    "results",
-    "appendix",
-    "other",
-  ]
-  print-metrics = true
-}
 reactome {
   inputs {
     pathways {
@@ -1476,6 +1412,71 @@ literature {
         format = ${common.output-format}
         path = ${literature.output}"/failedMatches"
       }
+    }
+
+    epmc {
+      uris {
+        ensembl = "https://www.ensembl.org/Homo_sapiens/Gene/Summary?db=core;g="
+        chembl = "https://www.ebi.ac.uk/chembl/compound_report_card/"
+        ontologies = "https://www.ebi.ac.uk/ols/ontologies/efo/terms?short_form="
+      }
+      input {
+        cooccurences = ${literature.processing.outputs.cooccurrences}
+      }
+      outputs {
+        epmc-cooccurrences {
+          format = ${common.output-format}
+          path = ${common.output}"/literature/epmcCooccurrences"
+        }
+        output {
+          format = "json" // to be consistent with the other Evidence files
+          path = ${common.input}"/evidence-files/epmc"
+          options = [
+            {k: "compression", v: "gzip"},
+          ]
+        }
+      }
+      excluded-target-terms = [
+        "TEC",
+        "TECS",
+        "Tec",
+        "tec",
+        "'",
+        "(",
+        ")",
+        "-",
+        "-S",
+        "S",
+        "S-",
+        "SS",
+        "SSS",
+        "Ss",
+        "Ss-",
+        "s",
+        "s-",
+        "ss",
+        "U3",
+        "U6",
+        "u6",
+        "SNORA70",
+        "U2",
+        "U8"
+      ]
+      # The following are the relevant sections for disease/target associations as described in PMID28587637
+      sections-of-interest = [
+        "title",
+        "abstract",
+        "intro",
+        "case",
+        "figure",
+        "table",
+        "discuss",
+        "concl",
+        "results",
+        "appendix",
+        "other",
+      ]
+      print-metrics = true
     }
   }
