+++ src/main/scala/io/opentargets/etl/Main.scala
@@ -8,10 +8,9 @@ import com.typesafe.scalalogging.LazyLogging
 import scala.util._
 import io.opentargets.etl.backend._
 import io.opentargets.etl.backend.drug.Drug
-import io.opentargets.etl.backend.epmc.Epmc
 import io.opentargets.etl.backend.evidence.Evidence
 import io.opentargets.etl.backend.graph.EtlDag
-import io.opentargets.etl.backend.literature.Literature
+import io.opentargets.etl.backend.literature.{Epmc, Literature}
 import io.opentargets.etl.backend.pharmacogenomics.Pharmacogenomics
 import io.opentargets.etl.backend.targetEngine.TargetEngine
 import io.opentargets.etl.common.GoogleStorageHelpers
