+++ src/main/scala/io/opentargets/etl/backend/Configuration.scala
@@ -376,7 +376,8 @@ object Configuration extends LazyLogging {
       common: LiteratureCommon,
       processing: LiteratureProcessing,
       embedding: LiteratureEmbedding,
-      vectors: LiteratureVectors
+      vectors: LiteratureVectors,
+      epmc: Epmc
   )
 
   case class TargetEngineInputs(targets: IOResourceConfig,
@@ -421,7 +422,6 @@ object Configuration extends LazyLogging {
       ebisearch: EBISearchSection,
       otarproject: OtarProjectSection,
       literature: LiteratureSection,
-      epmc: Epmc,
       targetEngine: TargetEngineSection
   )
 }
