+++ workflow/src/main/resources/reference.conf
@@ -32,8 +32,8 @@ labels {
   tool = "etl"
 }
 workflows = [
-  {name = "public", steps = ["disease", "epmc", "reactome", "pharmacogenomics", "expression", "go-step", "target", "interaction", "targetValidation", "evidence", "association", "associationOTF", "search", "drug", "knownDrug", "ebiSearch", "fda", "literature", "targetEngine"]}
-  {name = "private", steps = ["otar","disease", "epmc", "reactome", "pharmacogenomics", "expression", "go-step", "target", "interaction", "targetValidation", "evidence", "association", "associationOTF", "search", "drug", "knownDrug", "ebiSearch", "fda", "literature", "targetEngine"]}
+  {name = "public", steps = ["disease", "reactome", "pharmacogenomics", "expression", "go-step", "target", "interaction", "targetValidation", "evidence", "association", "associationOTF", "search", "drug", "knownDrug", "ebiSearch", "fda", "literature", "targetEngine"]}
+  {name = "private", steps = ["otar","disease", "reactome", "pharmacogenomics", "expression", "go-step", "target", "interaction", "targetValidation", "evidence", "association", "associationOTF", "search", "drug", "knownDrug", "ebiSearch", "fda", "literature", "targetEngine"]}
 ]
 
 existing-outputs {
@@ -66,10 +66,6 @@ jobs = [
     arg = "disease"
   },
   {
-    arg = "epmc"
-    deps = ["literature"]
-  },
-  {
     arg = "reactome"
   },
   {
@@ -97,7 +93,7 @@ jobs = [
   },
   {
     arg = "evidence"
-    deps = ["disease", "target", "epmc"]
+    deps = ["disease", "target", "literature"]
   },
   {
     arg = "association"
